requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
sqlalchemy
pymysql
langchain
mysql-connector-python
openai
azure-search-documents==11.4.0b6
azure-core 
azure-identity
azure-common
tiktoken